pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.101.0",
    "openai-whisper>=20250625",
    "python-dotenv>=1.1.1",
    "soundfile>=0.13.1",
    "streamlit>=1.48.1",
    "streamlit-webrtc==0.61.3",
    "torchaudio>=2.8.0",
    "webrtcvad>=2.0.10",
]
